Clean up of code in CleanUpBranch is complete.
Lol that felt good :). Apparently we're doing this lab again in lab 4 and 5 where the teacher said it was a good thing to get an idea of how the backend works. Hopefully comments should be able to explain that for the most part :P.

I'll be sure to merge this branch right away.